requirements.txt
flask
flask-cors
langchain
langchain-google-genai
langchain-openai
supabase
redis
python-dotenv
pydantic
openai
gunicorn
PyPDF2
pdfplumber
python-multipart
werkzeug
pymupdf
python-magic
loggings